refactor: extract ProjectGroup component from ProjectList
